% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsverse-package.R
\docType{package}
\name{phsverse-package}
\alias{phsverse}
\alias{phsverse-package}
\title{phsverse: Easily Install and Load Packages Developed in PHS}
\description{
This package is designed to make it easy to install and load multiple packages developed in PHS in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Public-Health-Scotland/phsverse}
  \item Report bugs at \url{https://github.com/Public-Health-Scotland/phsverse/issues}
}

}
\author{
\strong{Maintainer}: Tina Fu \email{Yuyan.Fu2@phs.scot}

Authors:
\itemize{
  \item Tina Fu \email{Yuyan.Fu2@phs.scot}
}

Other contributors:
\itemize{
  \item Public Health Scotland \email{phs.datascience@phs.scot} [copyright holder]
  \item James Hayes \email{James.Hayes2@phs.scot} (\href{https://orcid.org/0000-0002-5380-2029}{ORCID}) [contributor]
}

}
\keyword{internal}
