% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsverse.R
\docType{package}
\name{phsverse}
\alias{phsverse}
\title{\code{phsverse} package}
\description{
This package is designed to make it easy to install and load multiple packages developed in PHS in a single step.
}
\details{
See the README on
\href{https://github.com/Public-Health-Scotland/phsverse#readme}{GitHub}.
}
